Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,8 +1,8 @@
 #include<stdio.h>
 int main ()
 {
-    float x;
-    scanf("%f",&x);
+    int x;
+    scanf("%d",&x);
     printf("You entered: %.4f\n",x);
     return 0;
 }
